Apache: allow remote testing of SuEXEC configuration
Summary:
These tests are vital to run after each round of software upgrade,
as they ensure SuEXEC configuration is correct. As such, they should
be a part of the monitoring Jenkins job.

Plan is so to maintain a `qa` user account and populate with various
file, static HTML, regular CGI and PHP.

Tests also check autoindex directive in chmod 711 and 755 conditions.

Ref. T960

Reviewers: Sandlayth

Differential Revision: https://devcentral.nasqueron.org/D587