Hunt missing 502 pages for Docker services
Summary:
Each container on the Docker servers, when associated to an HTTP
host in the configuration, should serve a /502.html from rDII.

This change provides a class to read the rOPS configuration,
extract the domains names and check https://<host>/502.html.

When the page exists, it should return a 200 (502 is sent by the
web server on error, direct request of this page is a regular 200.)

Ref T114.

Test Plan: The test provided the new T114 list.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T114

Differential Revision: https://devcentral.nasqueron.org/D1967