Switch to Nasqueron codestyle
Summary:
The code used tabs on some files instead of spaces.

This change converts tabs to spaces and import phpcs Nasqueron rules.

An exemption is granted for line length, to allow tests to have
a verbose explanation about how to fix the issue. That will avoid
to rewrite them using HEREDOC. This only apply to tests, not lib
or the helper trait.

Test Plan: `phpcs`

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2589